feat: add update history and delete pokemon card
